backend/requirements.txt
asgiref==3.9.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
idna==3.10
itsdangerous==2.2.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.31.0
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extension